+++ src/root.zig
@@ -3,6 +3,6 @@
 
 const parser = @import("./core/parser.zig");
 
-
+pub const free = parser.free;
 pub const StaticJson = parser.Static;
 pub const DynamicJson = parser.Dynamic;