+++ CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 This project adheres to the [Semantic Version](https://semver.org/) guideline.
 
+## [v0.2.0] - 2025-02-12
+
+Supports automatic deep copy for parsed structure.
+
+### Added
+
+- Automatic Deep copy
+- Code Documentation
+- `parseInto()` on dynamic JSON
+- `free()` for deep copy structures
+
+### Changed
+
+- Working demos on `main.zig`
+
+### Fixed
+
+- Memory bugs on parsed structure
+
 ## [v0.1.0] - 2024-12-23
 
 Supports static and dynamic JSON data parsing and stringification.